Adiciona a documentação do comando de similaridade e corrige a comparação entre o valor da similaridade e a "nota de corte".

+++ documentstore_migracao/processing/compare_articles_sites.py
@@ -222,7 +222,7 @@ async def fetch_articles(session, pid, cut_off_mark, output_filepath):
     comp_data["found_text_classic"] = bool(comp_data["classic"])
     comp_data["found_text_new"] = bool(comp_data["new"])
 
-    if int(sim * 100) > cut_off_mark:
+    if (sim * 100) > cut_off_mark:
         del comp_data["classic"]
         del comp_data["new"]
